ms-usuarios/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
psycopg2-binary==2.9.9
flasgger==0.9.7.1